Refactor the to the required fomat
